refactor background.js and content.js
refactors the `background.js` and `content.js` files in order to improve the code structure and readability. additionally, it adds a new feature to the `background.js` file that allows for the fetching of the issue title from the openai api. in the `content.js` file, the issuediv is updated with the text from the api. finally, the `utils/index.js` file is updated to add the api token.

body:
* refactored `background.js` and `content.js` files
* added new feature to `background.js` to fetch issue title from openai api
* updated `content.js` to update issuediv with text from the api
* updated `utils/index.js` to add api token