adding the options.html file
adds the options.html file which contains an html form that allows the user to enter a personal token and clear their cache.

the options.html file contains a summary with information, a field for entering a personal token, and a button to clear the cache. the form also includes a link to generate a personal token and information about the safety of the token.