refactor scraped issue title
refactors the code in background.js to use a function when the key is `scraped_issue_title` and removes console.log from content.js. this ensures that the code is more efficient and organized.

in background.js, the `fetchsse` function is now awaited with the body of the message passed as an argument. in content.js, the `console.log` statement is removed and the `elementready` function is awaited before grabbing the titleelement and issuebodyelement. the textcontent is then used to set the value of the issuebodyelement.

body:
- background.js: refactor code to use a function when key is `scraped_issue_title` and await fetchsse with body of message passed as an argument
- content.js: remove console.log and await elementready before grabbing titleelement and issuebodyelement, then set value of issuebodyelement with textcontent