added click handler to the summarize button
adds a click handler to the summarize button that handles the scraping of issue title. it adds an `id` attribute to the `span` element and then adds a `observe` listener to trigger the click handler and start the scraping process.

the click handler is responsible for getting the issue title from the `titleelement` and sending it to the background script via `port.postmessage`. it also sets the `issuebodyelement` value to an empty string to reset the body field.