add interaction button to the new issue page
adds a button to the new issue page to allow users to interact with the page. the button is appended to the `sucessbtnpath` element on the page, which is selected using the `select-dom` library. the `elementready` function is used to ensure that the button is only added when the element is available on the page.

the changes made by this commit are as follows:
- import `select` from `select-dom` library
- add `sucessbtnpath` constant to store the path of the element where the button should be appended
- add `buttonready` constant and check if the button is already present on the page
- add `elementready` check for `sucessbtnpath` element
- append the `button()` to the `successbtn` element if the button is not already present on the page