add button to anchor on dom load
adds a button to the anchor element whenever it is loaded. it adds an `observe` event handler to the anchor element to detect when it is loaded and then appends a button to it. this allows the button to be dynamically added to the page whenever the anchor is loaded.

this commit adds the following changes:

- add `observe` event handler to the anchor element
- add a `buttonready` condition to check if the button has been added
- append button to the anchor element when `observe` event is triggered