update readme with firefox extension
updates the readme file with the link to firefox extension. it is a small change but it makes the readme more comprehensive and up to date.

the diff of this commit is as follows:
```diff
diff --git a/readme.md b/readme.md
index 3e6f64d98334..1ff2c3de04e5 100644
--- a/readme.md
+++ b/readme.md
@@ -8,7 +8,7 @@ it allows you to summarize a quick description about any new issue that you have

 ## install

-- [**firefox**]()
+- [**firefox** extension]()
 - [**chrome** extension]()

 ## usage
```