add auto-issue image .git preview
adds a new image to the readme.md file. the image is a gif of the auto-issue 5 feature. this image will help to better illustrate the feature and provide a visual aid for users.

```diff
diff --git a/readme.md b/readme.md
index d28704b1cb2b..e0946806eb51 100644
--- a/readme.md
+++ b/readme.md
@@ -6,6 +6,8 @@ it allows you to summarize a quick description about any new issue that you have

 ![auto-issue](https://user-images.githubusercontent.com/33763843/221444329-fee84a8e-64a8-4f4a-84c6-566fe7a8de77.gif)

+![autoissue5](https://user-images.githubusercontent.com/33763843/221567695-9ccc071a-fbf4-4e9a-99b4-df71c777caec.gif)
+
 ## install

 - [**firefox** extension]()
```