change readme gif
changes the gif image in the readme.md of the project.

the old image is replaced with a new one that better illustrates the capabilities of the project.

the diff of the commit is:
```
diff --git a/readme.md b/readme.md
index e0946806eb51..86f8a7f5d98f 100644
--- a/readme.md
+++ b/readme.md
@@ -4,8 +4,6 @@ chrome and firefox extension to automate the process of managing github issues u

 it allows you to summarize a quick description about any new issue that you have created or during the creation process. in other words this is useful as a quick start issue generator.

-![auto-issue](https://user-images.githubusercontent.com/33763843/221444329-fee84a8e-64a8-4f4a-84c6-566fe7a8de77.gif)
-
 ![autoissue5](https://user-images.githubusercontent.com/33763843/221567695-9ccc071a-fbf4-4e9a-99b4-df71c777caec.gif)

 ## install
```