added new funding file
adds the `.github/funding.yml` file to the repository, which provides a way to express and discover the financial sponsors of an open source project. this funding file includes a link to a github sponsors page where potential sponsors can easily learn more and offer financial support.

includes the following changes:

- added `.github/funding.yml` file
- updated link in `funding.yml` to include `github: falcucci/opensource`